Add minor msac optimizations
Skip the overhead of shifting in ones into the LSB in the common case,
that's only required in the EOB padding. In practice this means we
only have to invert bits once during the refill process instead of
twice in every call to msac functions.

Also make some improvements to the refill asm, mainly involving
keeping partially inserted bytes at the end instead of clearing them.