loopfilter: align Av1FilterLUT struct members
Fixes a bug where the Av1FilterLUT instance used in checkasm was not
aligned properly.

In theory, the first ALIGN macro should imply the latter alignments as well,
but I decided to mark all fields as explicitly aligned for clarity; and
because that's the precedent set in other headers.

Allows us to drop the ALIGN macro on the other usage of this struct.